refactor: create jsonschema and docstrings modules in utils
